login maintenance | register maintenance | main page maintenance | product-details page | review page | vote process | ...

+++ src/main/java/models/Person.java
@@ -1,14 +1,13 @@
 package models;
 
 public class Person {
-	private long id;
+	private int id;
 	private String email;
 	private String password;
-	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getEmail() {
@@ -23,5 +22,6 @@ public class Person {
 	public void setPassword(String password) {
 		this.password = password;
 	}
+	
 	
 }
